Rename scaffold to prototype
