+++ lib/ruby/signature.rb
@@ -29,8 +29,8 @@ require "ruby/signature/constant"
 require "ruby/signature/constant_table"
 require "ruby/signature/ast/comment"
 require "ruby/signature/writer"
-require "ruby/signature/scaffold/rbi"
-require "ruby/signature/scaffold/rb"
+require "ruby/signature/prototype/rbi"
+require "ruby/signature/prototype/rb"
 require "ruby/signature/environment_walker"
 
 require "ruby/signature/parser"
